+++ packages/terra-image/src/terra-dev-site/doc/image/Image.1.doc.jsx
@@ -10,6 +10,8 @@ import ImageSrc from '!raw-loader!../../../../src/Image';
 // Example Files
 import ImageDefault from '../example/ImageDefault';
 import ImageDefaultSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ImageDefault';
+import ImageFit from '../example/ImageFitTypes';
+import ImageFitSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ImageFitTypes';
 /* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
 const DocPage = () => (
@@ -22,6 +24,11 @@ const DocPage = () => (
         title: 'Default Image',
         example: <ImageDefault />,
         source: ImageDefaultSrc,
+      },
+      {
+        title: 'Toggle Image Fit',
+        example: <ImageFit />,
+        source: ImageFitSrc,
       },
     ]}
     propsTables={[