+++ packages/terra-form-input/CHANGELOG.md
@@ -9,6 +9,7 @@ Unreleased
 ### Changed
 * Minor dependency version bump
 * Updated errorIcon default prop from `Field.defaultProps.errorIcon` with `<IconError>` component.
+* updated package.json test scripts
 
 2.20.0 - (July 24, 2019)
 ------------------