+++ packages/terra-form-radio/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* updated package.json test scripts
 
 3.23.0 - (July 30, 2019)
 ------------------