+++ packages/terra-heading/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* updated package.json test scripts
 
 4.13.0 - (July 30, 2019)
 ------------------