+++ packages/terra-image/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* updated package.json test scripts
 
 3.14.0 - (July 30, 2019)
 ------------------
@@ -18,6 +20,10 @@ Unreleased
 ------------------
 ### Added
 * Added documentation on required peerDependencies
+* Added `Object-fit-images` polyfill for IE.
+* Added `object-fit` property in css.
+* Added prop Fit.
+* Added examples using Fit.
 
 ### Changed
 * Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`