+++ packages/terra-profile-image/CHANGELOG.md
@@ -3,6 +3,12 @@ Changelog
 
 Unreleased
 ----------
+### Updated
+* updated jest snapshots
+* updated to set `fit` to always be `cover` to handle non-square images.
+
+### Changed
+* updated package.json test scripts
 
 3.15.0 - (July 30, 2019)
 ------------------